Refactoring the registartion form.
